create module files
