+++ person.rb
@@ -34,33 +34,3 @@ class Person
         return ( parent_permission || age >= 18)
     end 
 end
-
-#Create Student Class
-class Student < Person
-    @classroom
-
-    #Create constructor
-    def initialize(classroom)
-        @classroom = classroom
-    end
-
-    #Creat play_hooky method
-    def play_hooky
-        "¯\(ツ)/¯"
-    end
-end
-
-#Create Teacher class
-class Teacher < Person
-    @specialization
-
-    #Create constructor
-    def initialize( specialization )
-        @specialization = specialization
-    end
-
-    #Create can_use_services? method to override can_use_services in Person class
-    def can_use_services?
-        return true
-    end
-end
